Gh 26383

+++ pkg/install/hypershift_test.go
@@ -19,7 +19,6 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/wait"
@@ -51,7 +50,7 @@ func initClient() ctrlClient.Client {
 
 func initDeployObj() *appsv1.Deployment {
 	return &appsv1.Deployment{
-		ObjectMeta: v1.ObjectMeta{
+		ObjectMeta: metav1.ObjectMeta{
 			Name:      util.HypershiftOperatorName,
 			Namespace: util.HypershiftOperatorNamespace,
 		},
@@ -72,7 +71,7 @@ func initDeployAddonImageDiffObj() *appsv1.Deployment {
 		util.HypershiftAddonAnnotationKey: util.AddonControllerName,
 	}
 	deploy.Spec.Template.Spec.Containers = []corev1.Container{
-		corev1.Container{Image: "testimage"},
+		{Image: "testimage"},
 	}
 	return deploy
 }
@@ -147,8 +146,8 @@ func (c *HypershiftTestCliExecutor) Execute(ctx context.Context, args []string) 
 			for _, tag := range ims.Spec.Tags {
 				if tag.Name == hsOperatorImage {
 					dp.Spec.Template.Spec.Containers[0].Image = tag.From.Name
+					break
 				}
-				break
 			}
 
 			break
@@ -414,7 +413,7 @@ func TestRunHypershiftInstall(t *testing.T) {
 	aCtrl.hubClient.Create(ctx, incompleteDp)
 
 	// No Spec in hypershift deployment operator - skip all operations
-	err := installHyperShiftOperator(t, ctx, aCtrl)
+	err := installHyperShiftOperator(t, ctx, aCtrl, true)
 	assert.Nil(t, err, "is nil if install HyperShift is successful")
 	aCtrl.hubClient.Delete(ctx, incompleteDp)
 
@@ -443,7 +442,7 @@ func TestRunHypershiftInstall(t *testing.T) {
 	aCtrl.hubClient.Create(ctx, dp)
 	defer aCtrl.hubClient.Delete(ctx, dp)
 
-	err = installHyperShiftOperator(t, ctx, aCtrl)
+	err = installHyperShiftOperator(t, ctx, aCtrl, true)
 	assert.Nil(t, err, "is nil if install HyperShift is successful")
 
 	// Check hypershift-operator-oidc-provider-s3-credentials secret exists
@@ -482,10 +481,11 @@ func TestRunHypershiftInstall(t *testing.T) {
 	ims := &imageapi.ImageStream{}
 	ims.Spec.Tags = tr
 	imb, err := yaml.Marshal(ims)
+	assert.Nil(t, err, "expected Marshal to succeed: %s", err)
 
 	// Run hypershift install again with image override
 	overrideCM := &corev1.ConfigMap{
-		ObjectMeta: v1.ObjectMeta{
+		ObjectMeta: metav1.ObjectMeta{
 			Name:      util.HypershiftDownstreamOverride,
 			Namespace: aCtrl.addonNamespace,
 		},
@@ -495,12 +495,12 @@ func TestRunHypershiftInstall(t *testing.T) {
 	aCtrl.hubClient.Create(ctx, overrideCM)
 	defer aCtrl.hubClient.Delete(ctx, overrideCM)
 
-	err = installHyperShiftOperator(t, ctx, aCtrl)
+	err = installHyperShiftOperator(t, ctx, aCtrl, true)
 	assert.Nil(t, err, "is nil if install HyperShift is sucessful")
 
 	// Run hypershift install again with image override using image upgrade configmap
 	imageUpgradeCM := &corev1.ConfigMap{
-		ObjectMeta: v1.ObjectMeta{
+		ObjectMeta: metav1.ObjectMeta{
 			Name:      util.HypershiftOverrideImagesCM,
 			Namespace: aCtrl.addonNamespace,
 		},
@@ -511,7 +511,7 @@ func TestRunHypershiftInstall(t *testing.T) {
 	err = aCtrl.hubClient.Create(ctx, imageUpgradeCM)
 	assert.Nil(t, err, "err nil when config map is created successfull")
 
-	err = installHyperShiftOperator(t, ctx, aCtrl)
+	err = installHyperShiftOperator(t, ctx, aCtrl, true)
 	assert.Nil(t, err, "is nil if install HyperShift is sucessful")
 
 	// Install hypershift job failed
@@ -527,7 +527,7 @@ func TestRunHypershiftInstall(t *testing.T) {
 	aCtrl.hubClient.Delete(ctx, pullSecret)
 	aCtrl.hubClient.Delete(ctx, hsPullSecret)
 
-	err = installHyperShiftOperator(t, ctx, aCtrl)
+	err = installHyperShiftOperator(t, ctx, aCtrl, true)
 	assert.Nil(t, err, "is nil if install HyperShift is sucessful")
 
 	err = aCtrl.spokeUncachedClient.Get(ctx, types.NamespacedName{Name: pullSecret.Name, Namespace: hypershiftOperatorKey.Namespace}, hsPullSecret)
@@ -535,7 +535,7 @@ func TestRunHypershiftInstall(t *testing.T) {
 
 	// Run hypershift install again with s3 bucket secret deleted
 	aCtrl.hubClient.Delete(ctx, bucketSecret)
-	err = installHyperShiftOperator(t, ctx, aCtrl)
+	err = installHyperShiftOperator(t, ctx, aCtrl, true)
 	assert.Nil(t, err, "is nil if install HyperShift is sucessful")
 
 	// Create hosted cluster
@@ -590,7 +590,7 @@ func TestReadDownstreamOverride(t *testing.T) {
 	assert.NotNil(t, err, "is not nil when read downstream image override fails")
 
 	overrideCM := &corev1.ConfigMap{
-		ObjectMeta: v1.ObjectMeta{
+		ObjectMeta: metav1.ObjectMeta{
 			Name:      util.HypershiftDownstreamOverride,
 			Namespace: aCtrl.addonNamespace,
 		},
@@ -722,6 +722,7 @@ func TestRunHypershiftInstallPrivateLinkExternalDNS(t *testing.T) {
 			"provider":      []byte(`aws`),
 			"credentials":   []byte(`private_secret`),
 			"domain-filter": []byte(`my.house.com`),
+			"txt-owner-id":  []byte(`the-owner`),
 		},
 	}
 
@@ -757,7 +758,8 @@ func TestRunHypershiftInstallPrivateLinkExternalDNS(t *testing.T) {
 	aCtrl.hubClient.Create(ctx, dp)
 	defer aCtrl.hubClient.Delete(ctx, dp)
 
-	err := installHyperShiftOperator(t, ctx, aCtrl)
+	err := installHyperShiftOperator(t, ctx, aCtrl, false)
+	defer deleteAllInstallJobs(ctx, aCtrl.spokeUncachedClient, aCtrl.addonNamespace)
 	assert.Nil(t, err, "is nil if install HyperShift is successful")
 
 	// Check hypershift-operator-oidc-provider-s3-credentials secret exists
@@ -792,6 +794,31 @@ func TestRunHypershiftInstallPrivateLinkExternalDNS(t *testing.T) {
 	err = aCtrl.spokeUncachedClient.Get(ctx, ctrlClient.ObjectKeyFromObject(edSecret), edSecret)
 	assert.Nil(t, err, "is nil when external dns secret is found")
 	assert.Equal(t, []byte(`private_secret`), edSecret.Data["credentials"], "the credentials should be equal if the copy was a success")
+
+	installJobList := &kbatch.JobList{}
+	err = aCtrl.spokeUncachedClient.List(ctx, installJobList)
+	if assert.Nil(t, err, "listing jobs should succeed: %s", err) {
+		if assert.Equal(t, 1, len(installJobList.Items), "there should be exactly one install job") {
+			installJob := installJobList.Items[0]
+			expectArgs := []string{
+				"--namespace", "hypershift",
+				"--oidc-storage-provider-s3-bucket-name", "my-bucket",
+				"--oidc-storage-provider-s3-region", "us-east-1",
+				"--oidc-storage-provider-s3-secret", "hypershift-operator-oidc-provider-s3-credentials",
+				"--aws-private-secret", "hypershift-operator-private-link-credentials",
+				"--aws-private-region", "us-east-1",
+				"--private-platform", "AWS",
+				"--external-dns-secret", "hypershift-operator-external-dns-credentials",
+				"--external-dns-domain-filter", "my.house.com",
+				"--external-dns-provider", "aws",
+				"--external-dns-txt-owner-id", "the-owner",
+				"--enable-uwm-telemetry-remote-write",
+				"--platform-monitoring", "OperatorOnly",
+				"--hypershift-image", "my-test-image",
+			}
+			assert.Equal(t, expectArgs, installJob.Spec.Template.Spec.Containers[0].Args, "mismatched container arguments")
+		}
+	}
 
 	// Cleanup
 	err = aCtrl.RunHypershiftCmdWithRetries(ctx, 3, time.Second*10, aCtrl.RunHypershiftCleanup)
@@ -862,11 +889,13 @@ func getHostedCluster(hcNN types.NamespacedName) *hyperv1alpha1.HostedCluster {
 	return hc
 }
 
-func installHyperShiftOperator(t *testing.T, ctx context.Context, aCtrl *UpgradeController) error {
+func installHyperShiftOperator(t *testing.T, ctx context.Context, aCtrl *UpgradeController, deleteJobs bool) error {
 	go updateHsInstallJobToSucceeded(ctx, aCtrl.spokeUncachedClient, aCtrl.addonNamespace)
 	err := aCtrl.RunHypershiftInstall(ctx)
-	if err := deleteAllInstallJobs(ctx, aCtrl.spokeUncachedClient, aCtrl.addonNamespace); err != nil {
-		t.Errorf("error cleaning up HyperShift install jobs: %s", err.Error())
+	if deleteJobs {
+		if err := deleteAllInstallJobs(ctx, aCtrl.spokeUncachedClient, aCtrl.addonNamespace); err != nil {
+			t.Errorf("error cleaning up HyperShift install jobs: %s", err.Error())
+		}
 	}
 
 	return err
